# Do WBL mmap marker replay concurrently